requirements.txt
bs4
pdbpp
pymongo
python-dotenv
requests
selenium>=3.141.0
